finished authentication ande secured routes

+++ app/Http/Controllers/ListingController.php
@@ -50,6 +50,8 @@ class ListingController extends Controller
         if($request->hasFile('logo')) {
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
+
+        $formFields['user_id'] = auth()->id();
         
 
         Listing::create($formFields);
@@ -65,6 +67,11 @@ class ListingController extends Controller
     // update listing data
     public function update(Request $request, Listing $listing) 
     {
+        // make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, "Unauthorized action");
+        }
+
         $formFields = $request->validate([
             'title'   => 'required',
             'company' => ['required'],
@@ -87,7 +94,17 @@ class ListingController extends Controller
 
     public function destroy(Listing $listing)
     {
+        // make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, "Unauthorized action");
+        }
+        
         $listing->delete();  
         return redirect('/')->with('message', 'Listing successfully deleted!');
+    }
+
+    public function manage()
+    {
+        return view('listings.manage', ['listings' => auth()->user()->listing()->get()]);
     }
 }
